+++ wayland-server/build.rs
@@ -11,5 +11,10 @@ fn main() {
     let out_dir = Path::new(&out_dir_str);
 
     println!("cargo:rerun-if-changed={}", protocol_file);
-    generate_code(protocol_file, out_dir.join("wayland_api.rs"), Side::Server);
+    generate_code_with_destructor_events(
+        protocol_file,
+        out_dir.join("wayland_api.rs"),
+        Side::Server,
+        &[("wl_callback", "done")],
+    );
 }